Move DrawTile method to new CanvasArena class
