+++ src/razor-core/DevilDaggersInfo.Razor.Core.Canvas/Canvas2d.cs
@@ -536,8 +536,4 @@ public class Canvas2d : Canvas
 
 	public void Restore()
 		=> Invoke("c2d.restore");
-
-	// TODO: Move.
-	public void DrawTile(int x, int y, int r, int g, int b, float tileSize)
-		=> Invoke("c2d.drawTile", x, y, r, g, b, tileSize);
 }
